backend/requirements/requirements.txt
click==6.7
defusedxml==0.5.0
Django==1.10.5
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
oauthlib==2.0.2
PyJWT==1.5.0
python-dotenv==0.6.4
python3-openid==3.1.0
requests==2.14.2
requests-oauthlib==0.8.0
rest-social-auth==1.1.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
